Basically it works. Java layer now supports the new c layer interfaces. Some pieces of code were rewritten to make them simplier in use. Also I definetely want to do something with listeners.

+++ CatEyeB.BrightnessStageOperation/com/cateye/stageoperations/brightness/BrightnessStageOperation.java
@@ -3,19 +3,24 @@ package com.cateye.stageoperations.brightness;
 import com.cateye.core.ArgumentOutOfRangeException;
 import com.cateye.core.stage.StageOperation;
 
-public class BrightnessStageOperation extends StageOperation {
-	private float brightness;
-
-	public float getBrightness() {
+public class BrightnessStageOperation extends StageOperation
+{
+	private double brightness;
+	
+	public double getBrightness()
+	{
 		return brightness;
 	}
-
-	public void setBrightness(float brightness) {
-		if (brightness < 0.01f || brightness > 100f) {
+	
+	public void setBrightness(double brightness)
+	{
+		if (brightness < 0.01d || brightness > 100d)
+		{
 			throw new ArgumentOutOfRangeException("brightness");
 		}
-
-		if ((brightness - this.brightness) >= Float.MIN_NORMAL) {
+		
+		if ((brightness - this.brightness) >= Double.MIN_NORMAL)
+		{
 			this.brightness = brightness;
 			fireOnPropertyChanged("brightness", brightness);
 		}
